Removed dxpy package from most classes & refactored for more obvious flow
