+++ collapsevariants/collapse_variants.py
@@ -10,14 +10,15 @@ from pathlib import Path
 from typing import Dict
 
 import dxpy
-from general_utilities.association_resources import generate_linked_dx_file
+from general_utilities.import_utils.file_handlers.dnanexus_utilities import generate_linked_dx_file
 from general_utilities.mrc_logger import MRCLogger
 
-from collapsevariants.collapse_logger import CollapseLOGGER
-from collapsevariants.ingest_data import IngestData
-from collapsevariants.parallelization_wrappers import generate_generic_masks, generate_snp_or_gene_masks, \
-    generate_genotype_matrices, update_log_file
-from collapsevariants.snp_list_generator import SNPListGenerator
+from utilities.collapse_logger import CollapseLOGGER
+from utilities.ingest_data import IngestData
+from utilities.parallelization_wrappers import generate_generic_masks, generate_snp_or_gene_masks, \
+    update_log_file
+from genotype_matrix.genotype_matrix import generate_genotype_matrices
+from collapsevariants.snp_list_generator.snp_list_generator import SNPListGenerator
 
 # Set up the system logger – this is NOT the same as LOG_FILE below that records info about the filtering itself
 LOGGER = MRCLogger().get_logger()
@@ -57,12 +58,12 @@ def main(filtering_expression: str, snplist: dict, genelist: dict, output_prefix
 
     # 2. Generate a list of ALL variants genome-wide that we want to retain:
     LOGGER.info('Filtering variants according to provided inputs...')
-    snp_list_generator = SNPListGenerator(ingested_data.vep_dict, ingested_data.filtering_expression,
+    snp_list_generator = SNPListGenerator(ingested_data.bgen_dict, ingested_data.filtering_expression,
                                           ingested_data.gene_list_path, ingested_data.snp_list_path, log_file)
 
     # 3. Generate genotype matrices for each bgen file:
     LOGGER.info('Generating genotype matrices for each bgen file...')
-    genotype_index = generate_genotype_matrices(snp_list_generator.genes, ingested_data.bgen_index)
+    genotype_index = generate_genotype_matrices(snp_list_generator.genes, ingested_data.bgen_dict)
 
     # 4. Write information about collapsing to the log file
     LOGGER.info('Updating log file with per-sample and per-ENST totals...')
@@ -85,7 +86,7 @@ def main(filtering_expression: str, snplist: dict, genelist: dict, output_prefix
                                               output_prefix)
 
     LOGGER.info('Closing LOG file...')
-    linked_log_file = log_file.close_and_upload()
+    log_file.close()
 
     # Here we are taking all the files generated by the various functions above and adding them to a single tar
     # to enable easy output. The only output of this applet is thus a single .tar.gz file per VCF file
@@ -100,7 +101,7 @@ def main(filtering_expression: str, snplist: dict, genelist: dict, output_prefix
 
     # Set output
     output = {'output_tarball': dxpy.dxlink(generate_linked_dx_file(output_tarball)),
-              'log_file': dxpy.dxlink(linked_log_file)}
+              'log_file': dxpy.dxlink(generate_linked_dx_file(log_file.get_path()))}
 
     return output
