+++ collapsevariants/utilities/ingest_data.py
@@ -1,27 +1,26 @@
 import csv
 from pathlib import Path
-from typing import TypedDict, Dict, Optional, IO, Tuple
+from typing import TypedDict, Dict, Optional, Tuple
 
-import dxpy
-from general_utilities.association_resources import download_dxfile_by_name
+from general_utilities.import_utils.file_handlers.input_file_handler import InputFileHandler
 from general_utilities.mrc_logger import MRCLogger
 
-from collapsevariants.collapse_utils import get_sample_ids
+from utilities.collapse_utils import get_sample_ids
 
 
 class BGENIndex(TypedDict):
     """A TypedDict containing information on a downloaded bgen file to make for more obvious return types
 
-    :cvar index: A DXFile representation of the .bgen.bgi index file
-    :cvar sample: A DXFile representation of the bgen .sample file
-    :cvar bgen: A DXFile representation of the .bgen genetic data file
-    :cvar vep: A DXFile representation of the annotation .tsv.gz file generated by VEP
+    :cvar index: A InputFileHandler representation of the .bgen.bgi index file
+    :cvar sample: A InputFileHandler representation of the bgen .sample file
+    :cvar bgen: A InputFileHandler representation of the .bgen genetic data file
+    :cvar vep: A InputFileHandler representation of the annotation .tsv.gz file generated by VEP
     """
 
-    index: str
-    sample: str
-    bgen: str
-    vep: str
+    index: InputFileHandler
+    sample: InputFileHandler
+    bgen: InputFileHandler
+    vep: InputFileHandler
 
 
 class IngestData:
@@ -34,8 +33,8 @@ class IngestData:
 
     :param filtering_expression: A string filtering expression to filter variants (must be compatible with pandas.query())
     :param bgen_index: A file containing information of bgen files to collapse on
-    :param snp_list: A DXFile containing a list of varIDs to use as a custom mask
-    :param gene_list: A DXFile containing a list of gene symbols to collapse into a custom mask
+    :param snp_list: A file containing a list of varIDs to use as a custom mask
+    :param gene_list: A file containing a list of gene symbols to collapse into a custom mask
     """
 
     def __init__(self, bgen_index: dict, filtering_expression: str, snp_list: Optional[dict],
@@ -47,8 +46,7 @@ class IngestData:
         self.filtering_expression = filtering_expression
 
         self.sample_ids = []
-        self.bgen_index = self._ingest_bgen_index(bgen_index)
-        self.vep_dict = self._open_vep_filepaths()
+        self.bgen_dict = self._ingest_bgen_index(bgen_index)
         self.snp_list_path = self._define_filter_list(snp_list)
         self.gene_list_path = self._define_filter_list(gene_list)
 
@@ -67,58 +65,45 @@ class IngestData:
         """
 
         bgen_dict = {}
-        bgen_index = download_dxfile_by_name(bgen_index)
+        bgen_index = InputFileHandler(bgen_index, download_now=True)
         samples_collected = False
 
         # and load it into a dict:
-        with bgen_index.open('r') as bgen_reader:
+        with bgen_index.get_file_handle().open('r') as bgen_reader:
             bgen_index_csv = csv.DictReader(bgen_reader, delimiter='\t')
 
             for batch in bgen_index_csv:
 
-                bgen_dict[batch['prefix']] = {'index': batch['bgen_index_dxid'], 'sample': batch['sample_dxid'],
-                                              'bgen': batch['bgen_dxid'], 'vep': batch['vep_dxid']}
+                bgen_dict[batch['prefix']] = {'index': InputFileHandler(batch['bgen_index_id']),
+                                              'sample': InputFileHandler(batch['sample_id']),
+                                              'bgen': InputFileHandler(batch['bgen_id']),
+                                              'vep': InputFileHandler(batch['vep_id'])}
 
                 # Collect sample IDs from the 1st bgen encountered. This is done on the premise that all bgen files
                 # have the same sample IDs...
                 if samples_collected is False:
-                    sample_file = download_dxfile_by_name(batch['sample_dxid'], print_status=False)
-                    self.sample_ids = get_sample_ids(sample_file)
+
+                    sample_file = bgen_dict[batch['prefix']]['sample']
+                    sample_path = sample_file.get_file_handle()
+                    self.sample_ids = get_sample_ids(sample_path)
                     samples_collected = True
-                    sample_file.unlink()  # Make sure to delete to avoid conflict with later download
+                    sample_file.get_file_handle().unlink()  # Make sure to delete to avoid conflict with later download
 
         if len(self.sample_ids) == 0:
             raise ValueError('No sample IDs found in the bgen files. Please check the sample files and try again.')
 
         return bgen_dict
 
-    def _open_vep_filepaths(self) -> Dict[str, IO]:
-        """A convenience method to open an IO stream all VEP files for the bgen files in the bgen_index.
-
-        This method was created to allow for easier testing of downstream methods that require the VEP files to be open
-        rather than opened as part of the downstream method. This is because these files are directly streamed from the
-        DNANexus platform and are not stored locally. This means that they are not available for testing on a local
-        machine.
-
-        :return: A dictionary with keys of bgen_prefix and values of the VEP file IO stream.
-        """
-
-        vep_dict = {}
-        for bgen_prefix, bgen_info in self.bgen_index.items():
-            vep_dict[bgen_prefix] = dxpy.open_dxfile(bgen_info['vep'], mode='rb')
-
-        return vep_dict
-
     @staticmethod
-    def _define_filter_list(filtering_list: dict) -> Optional[Path]:
+    def _define_filter_list(filtering_list: dict) -> Optional[InputFileHandler]:
         """Download a filtering list file (if provided)
 
-        :param filtering_list: A DXFile ID pointing to some filtering file (likely Gene / SNP-based) on the RAP
+        :param filtering_list: A InputFileHandler pointing to some filtering file (likely Gene / SNP-based) on the RAP
         :return: A Path object pointing to the downloaded file, if found, otherwise None
         """
 
         if filtering_list:
-            return download_dxfile_by_name(filtering_list)
+            return InputFileHandler(filtering_list)
         else:
             return None
 
@@ -135,8 +120,8 @@ def download_bgen(chrom_bgen_index: BGENIndex) -> Tuple[Path, Path, Path]:
     """
 
     # Download the requisite files for this chromosome according to the index dict:
-    bgen_path = download_dxfile_by_name(chrom_bgen_index['bgen'], print_status=False)
-    index_path = download_dxfile_by_name(chrom_bgen_index['index'], print_status=False)
-    sample_path = download_dxfile_by_name(chrom_bgen_index['sample'], print_status=False)
+    bgen_path = chrom_bgen_index['bgen'].get_file_handle()
+    index_path = chrom_bgen_index['index'].get_file_handle()
+    sample_path = chrom_bgen_index['sample'].get_file_handle()
 
     return bgen_path, index_path, sample_path
